populated with acceptable data. schema.sql added. access control implemented

+++ controllers/ApikeyController.php
@@ -8,6 +8,8 @@ use app\models\searches\ApikeySearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use yii\filters\AccessControl;
+use yii\web\ForbiddenHttpException;
 
 /**
  * ApikeyController implements the CRUD actions for Apikey model.
@@ -26,15 +28,34 @@ class ApikeyController extends Controller
                     'delete' => ['POST'],
                 ],
             ],
+            'access' => [
+                'class' => AccessControl::className(),
+                'only' => ['home','index','view','create','update','delete'],
+                'rules' => [
+                    [
+                        'allow' => false,
+                        'roles' => ['?'],
+                    ],
+                    [
+                        'allow' => true,
+                        'actions' => ['index','view','create','update','delete'],
+                        'roles' => ['@'],
+                    ],
+                ],
+            ],
         ];
     }
 
     /**
      * Lists all Apikey models.
      * @return mixed
+     * @throws ForbiddenHttpException
      */
     public function actionIndex()
     {
+        if(Yii::$app->user->identity->role->name !== "SA"){
+            throw new ForbiddenHttpException('You are not allowed to enter this section.');
+        }
         $searchModel = new ApikeySearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -49,9 +70,13 @@ class ApikeyController extends Controller
      * @param integer $id
      * @return mixed
      * @throws NotFoundHttpException if the model cannot be found
+     * @throws ForbiddenHttpException
      */
     public function actionView($id)
     {
+        if(Yii::$app->user->identity->role->name !== "SA"){
+            throw new ForbiddenHttpException('You are not allowed to enter this section.');
+        }
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
@@ -61,9 +86,13 @@ class ApikeyController extends Controller
      * Creates a new Apikey model.
      * If creation is successful, the browser will be redirected to the 'view' page.
      * @return mixed
+     * @throws ForbiddenHttpException
      */
     public function actionCreate()
     {
+        if(Yii::$app->user->identity->role->name !== "SA"){
+            throw new ForbiddenHttpException('You are not allowed to enter this section.');
+        }
         $model = new Apikey();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -81,9 +110,13 @@ class ApikeyController extends Controller
      * @param integer $id
      * @return mixed
      * @throws NotFoundHttpException if the model cannot be found
+     * @throws ForbiddenHttpException
      */
     public function actionUpdate($id)
     {
+        if(Yii::$app->user->identity->role->name !== "SA"){
+            throw new ForbiddenHttpException('You are not allowed to enter this section.');
+        }
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -101,9 +134,13 @@ class ApikeyController extends Controller
      * @param integer $id
      * @return mixed
      * @throws NotFoundHttpException if the model cannot be found
+     * @throws ForbiddenHttpException
      */
     public function actionDelete($id)
     {
+        if(Yii::$app->user->identity->role->name !== "SA"){
+            throw new ForbiddenHttpException('You are not allowed to enter this section.');
+        }
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
